fix adding posts to homepage Relates #30
