+++ public/pages/createpost.html
@@ -40,8 +40,8 @@
             </div>
         </div>
     </section>
-    <script src="../javascript/createpost.js"></script>
     <script src="../javascript/helpers.js"></script>
+    <script src="../javascript/createpost.js"></script>
 
 </body>
